0.3.1 Huge number of changes. This feature was all around generating summary data tables. All the study functions were changes.
Introduced the concept of creating an install script. This seems to work OK.

Created function to load summary-data and a shell script.

Update the load script to include building the summary table.

Updated data for analytic tables and functions. This was a major part of this feture.
Now everything pretty much goes out of these tables and functions.